# wongnok-service/cmd/server/docs/swagger.yaml
definitions:
...
  dto.FavoriteResponse:
    properties:
      foodRecipeID:
        type: integer
      id:
        type: integer
      userID:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: This is an wongnok server.
  title: Wongnok API
  version: "1.0"
paths:
...
  /api/v1/food-recipes/{id}/favorites:
    delete:
      consumes:
      - application/json
      description: delete favorite by fooderecipeid
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FavoriteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a favorite
      tags:
      - favorite
    get:
      consumes:
      - application/json
      description: Get all favorites
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FavoriteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get favorites
      tags:
      - favorite
    post:
      consumes:
      - application/json
      description: Create a new favorite for a food recipe by userID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FavoriteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a favorite
      tags:
      - favorite
...
  /api/v1/food-recipes/favorites:
    get:
      consumes:
      - application/json
      description: Get all favorites
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Items per page
        in: query
        name: limit
        required: true
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FavoriteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: GetByuser favorites
      tags:
      - favorite
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"